requirements.txt
discord.py==2.3.2
requests==2.31.0
fastapi==0.110.0
uvicorn==0.29.0
websockets==12.0
centrifuge-python==0.3.0
requests-oauthlib==2.0.0